// src/main/model/entity/User.ts
import { Column, DataType, HasOne, Model, Table } from "sequelize-typescript"
import Customers from "./Customer"
import Stores from "./Store"
import { Sequelize } from "sequelize"

export interface UserAttributes {
    id: string
    email: string
    username: string
    password: string
    createdAt: Date
    createdBy: string
    updatedAt: Date
    updatedBy: string
    deletedAt: Date | null
    deletedBy: string | null
}

type UserAutoGeneratedAttributes = "id" | "createdAt" | "updatedAt" | "deletedAt"
type UserOptionalAttributes = "createdBy" | "updatedBy" | "deletedBy"
export type UserCreationAttributes = Omit<UserAttributes, UserAutoGeneratedAttributes | UserOptionalAttributes> & Partial<Pick<UserAttributes, UserOptionalAttributes>>

@Table({
    freezeTableName: true,
    tableName: "user",
    paranoid: true,
    underscored: true,
    schema: "public"
})
class User extends Model<UserAttributes | UserCreationAttributes> implements UserAttributes {
    @Column({
        allowNull: false,
        type: DataType.UUID,
        primaryKey: true,
        defaultValue: Sequelize.literal("uuid_generate_v7()")
    })
    id!: string

    @Column({
        allowNull: false,
        type: DataType.STRING,
        unique: true
    })
    email!: string

    @Column({
        allowNull: false,
        type: DataType.STRING,
        unique: true
    })
    username!: string

    @Column({
        allowNull: false,
        type: DataType.STRING
    })
    password!: string

    @Column({
        allowNull: false,
        type: DataType.DATE,
        defaultValue: Sequelize.literal("now()")
    })
    createdAt!: Date

    @Column({
        allowNull: false,
        type: DataType.STRING(128),
        defaultValue: "System",
    })
    createdBy!: string

    @Column({
        allowNull: false,
        type: DataType.DATE,
        defaultValue: Sequelize.literal("now()")
    })
    updatedAt!: Date

    @Column({
        allowNull: false,
        type: DataType.STRING(128),
        defaultValue: "System",
    })
    updatedBy!: string

    @Column({
        allowNull: true,
        type: DataType.DATE
    })
    deletedAt!: Date | null

    @Column({
        allowNull: true,
        type: DataType.STRING(128)
    })
    deletedBy!: string | null

    @HasOne(() => Customers, { foreignKey: "userId", sourceKey: "id", onDelete: "CASCADE" })
    customer!: Customers

    @HasOne(() => Stores, { foreignKey: "userId", sourceKey: "id", onDelete: "CASCADE" })
    store!: Stores
}

export default User